requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cloudpickle==3.1.1
cryptography==45.0.7
dataclasses-json==0.6.7
datauri==1.0.0
distro==1.9.0
easygui==0.98.3
eval_type_backport==0.2.2
feedparser==6.0.12
filetype==1.2.0
frozenlist==1.7.0
fsspec==2024.12.0
google==3.0.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
invoke==2.2.0
jiter==0.10.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-anthropic==0.3.20
langchain-community==0.3.29
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-google-genai==2.1.10
langchain-mistralai==0.2.11
langchain-neo4j==0.5.0
langchain-openai==0.3.33
langchain-perplexity==0.1.2
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.27
marshmallow==3.26.1
mistralai==1.9.10
mkl-service==2.4.0
multidict==6.6.4
mypy_extensions==1.1.0
neo4j==5.28.2
neo4j-graphrag==1.10.0
ollama==0.5.3
openai==1.107.2
orjson==3.11.3
ormsgpack==1.10.0
pandas==2.3.2
pdf2doi==1.7
pdfminer.six==20221105
pdftitle==0.11
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMuPDF==1.26.4
pypdf==6.0.0
PyPDF2==2.0.0
pyperclip==1.9.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scipy==1.16.2
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlite-vec==0.1.6
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.0
tqdm==4.67.1
transformers==4.56.1
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
